+++ src/aws/pci-dss-3.2.1/index.ts
@@ -4,4 +4,5 @@ export default {
   provider: 'aws',
   entity: 'PCI',
   rules: PolicyPacksRules,
+  extraFields: ['arn', 'accountId'],
 }